Committing lots of work before starting cleanup

+++ mhash_md5.h
@@ -3,9 +3,10 @@
 
 #include <stdlib.h>
 #include <stdio.h>
+#include <string.h>
 #include <inttypes.h>
 
-uint8_t* mhash_md5_buf(uint8_t*, size_t);
+uint8_t* mhash_md5_buf(uint8_t*, uint32_t);
 uint8_t* mhash_md5_file(FILE*);
 
 char* mhash_md5_str(uint8_t*);
